requirements.txt
amqp==5.0.6
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.6
django-celery-results==2.2.0
djangorestframework==3.12.4
idna==3.2
kombu==5.1.0
prompt-toolkit==3.0.20
pycodestyle==2.7.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
